requirements.txt
openai>=1.12.0
chromadb>=0.4.24
numpy>=1.24.0
python-dotenv>=1.0.0 